--- .babelrc
@@ -1,3 +1,0 @@
-{
-  "stage":  1
-}
