+++ .gitignore
@@ -2,3 +2,5 @@
 /taskcluster-auth.conf.json
 /user-config.yml
 /npm-debug.log*
+/lib/
+/.test/
