+++ bin/load-test.js
@@ -2,7 +2,7 @@ var debug         = require('debug')('queue:bin:load-test');
 var base          = require('taskcluster-base');
 var Promise       = require('promise');
 var _             = require('lodash');
-var v1            = require('../auth/v1');
+var v1            = require('../lib/v1');
 var https         = require('https');
 var http          = require('http');
 var hawk          = require('hawk');
