+++ src/azure.js
