+++ src/dfa.js
