+++ src/scoperesolver.js
