+++ src/sentrymanager.js
