+++ src/signaturevalidator.js
