+++ test/mocha.opts
@@ -1,4 +1,3 @@
 --ui tdd
 --timeout 30s
 --reporter spec
---compilers js:babel/register
